Additional for_each functionality
One can now iterate from the middle of an ordered_container using :for_each( collection, 5, ... ), where 5 is the starting index. This is enabled only for regular collections and not for ordered_collection. The reasons is that the ordered collection performance will be suboptimal. We may want to add that when we have established a standard of adding "slow" functions.